+++ packages/on-error-resume-next/src/index.ts
@@ -1,3 +1,5 @@
+import isPromise from './private/isPromise';
+
 export function onErrorResumeNext<T extends () => U, U = unknown>(fn: T, context?: undefined): U | undefined;
 export function onErrorResumeNext<T extends (this: V) => U, U = unknown, V = undefined>(
   fn: T,
@@ -8,9 +10,17 @@ export function onErrorResumeNext<T extends (this: V) => U, U = unknown, V = und
   fn: T,
   context: V
 ): U | undefined {
+  let result: U;
+
   try {
-    return fn.call(context);
+    result = fn.call(context);
   } catch {
-    return;
+    return undefined;
   }
+
+  if (isPromise(result)) {
+    throw new Error('Promise is not supported, please use "on-error-resume-next/async" instead.');
+  }
+
+  return result;
 }